refactor: order와 table 의존성 단방향으로 분리
