+++ src/main/java/kitchenpos/table/domain/OrderTableStatusValidator.java
@@ -1,0 +1,9 @@
+package kitchenpos.table.domain;
+
+import java.util.List;
+
+public interface OrderTableStatusValidator {
+    void validateOrderStatus(Long orderTableId);
+
+    void validateOrderTablesStatus(List<OrderTable> orderTables);
+}
